Updated client, renamed /open to /reopen
