+++ prow/plugins/BUILD
@@ -52,8 +52,8 @@ filegroup(
         "//prow/plugins/heart:all-srcs",
         "//prow/plugins/label:all-srcs",
         "//prow/plugins/lgtm:all-srcs",
-        "//prow/plugins/open:all-srcs",
         "//prow/plugins/releasenote:all-srcs",
+        "//prow/plugins/reopen:all-srcs",
         "//prow/plugins/trigger:all-srcs",
         "//prow/plugins/yuks:all-srcs",
     ],
